vm: fix some size_t/uint64_t issues (7)

+++ src/cshared/stream.cpp
@@ -173,7 +173,7 @@ byte *Stream::data() const
     return buffer.data + _position;
 }
 
-size_t Stream::remaining() const
+uint64_t Stream::remaining() const
 {
     if (buffer.length <= _position)
         return 0;
